Finished converting input and binding output to view

+++ DeathTimer/Models/WeightedInput.cs
@@ -9,15 +9,15 @@ namespace DeathTimer
     //This class contains weighted values of input parameters in application
     class WeightedInput
     {
-        public float Gender { get; set; }
-        public float Age { get; set; }
-        public float Height { get; set; }
-        public float Weight { get; set; }
-        public float Stress { get; set; }
-        public float Calories { get; set; }
-        public float Alcohol { get; set; }
-        public float Cigarettes { get; set; }
-        public float Job { get; set; }     
-        public float Activity { get; set; }
+        public int Gender { get; set; }
+        public int Age { get; set; }
+        public int Height { get; set; }
+        public int Weight { get; set; }
+        public int Stress { get; set; }
+        public int Calories { get; set; }
+        public int Alcohol { get; set; }
+        public int Cigarettes { get; set; }
+        public int Job { get; set; }     
+        public int Activity { get; set; }
     }
 }
